samosh/requirements.txt
python-telegram-bot==13.15
SQLAlchemy==2.0.22
psycopg2==2.9.8
python-dotenv==1.0.0